f3: add Path to the settings
The mapping between the IDs of the remote forge and the local Forgejo
instance must be preserved somewhere. Since the preferred method for
storing a F3 archive is a repository, having a directory where it is
cloned is convenient.

By default it is in the f3 subdirectory of the data path.