feat: Allow package overwrite
Currently, if you try to upload a package to Forgejo that already exists
with the same name/version, you will get an HTTP 409 conflict error. So,
if you want to edit/overwrite a package, you must first delete it.

This isn't always desireable, so let's add a way to opt-in to allowing
package overwrites with a new config variable:

[packages]
ALLOW_OVERWRITE = true