fix(ui): don't use `ui comments` in code comments
The only other usage of .comments I found in the UI is in the `new issue` form: here it places avatar and issue form correctly with `float`.. even issue view doesn't use it and uses another implementation where avatar is shifted `left` with negative value.

In code comments `.comments` was mostly causing bugs:
* horizontal avatar misglignment
* lack of rounded corners on avatar image

This commit breaks alignment of username and content, it will be fixed in next commits.