feat: always publish the link to the commit status
The current implementation is limited to self-hosted actions and buggy
as soon as multiple repos are involved, like for the homepage (because
each permission must be fetched individually).

Ideally this feature should work for all kind of status (with some
setting indicating which collaborator can access with status).

Cf https://codeberg.org/forgejo/forgejo/pulls/8152