fix: do not ignore automerge while a PR is checking for conflicts
Automerge can be ignored when the following race happens:

* Conflict check is happening on a repository and
  `pr.Status = issues_model.PullRequestStatusChecking` for all open pull
  requests (this happens every time a pull request is merged).
* While the conflict check is ongoing, an event (Forgejo Actions being
  successful for instance) happens and and `StartPRCheckAndAutoMerge*` is called.
* Because `pr.CanAutoMerge()` is false, the pull request is not
  selected and not added to the automerge queue.
* When the conflict check completes and `pr.CanAutoMerge()` becomes
  true, there no longer is a task in the auto merge queue and the
  auto merge does not happen.

This is fixed by adding a task to the auto merge queue when the
conflict check for a pull request completes. This is done when
the mutx protecting the conflict check task is released to prevent
a deadlock when a synchronous queues are used in the following
situation:

* the conflict check task finds the pull request is mergeable
* it schedules the auto merge tasks that finds it must be merged
* merging concludes with scheduling a conflict check task

Avoid an extra loop where a conflict check task queues an auto merge
task that will schedule a conflict check task if the pull request can
be merged. The auto merge row is removed from the database before
merging. It would otherwise be removed after the merge commit is
received via the git hook which happens asynchronously and can lead to
a race.

StartPRCheckAndAutoMerge is modified to re-use HeadCommitID when
available, such as when called after a pull request conflict
check.

Closes forgejo/forgejo#3470