chore: do not require empty fixtures to clean tables
- A mistake that is often made is to not put a empty fixture for tables
that gets populated in a test and should be cleaned up between runs. The
CI does not detect such issues as these never arise on the first run of the
test suite and are only noticed when developers run a test for a second
time, unless you are aware of this behavior (which very few do as this
is not documented and pure folk knowledge) can end up in chasing a bug
that does not exist for hours. forgejo/forgejo#7041,
forgejo/forgejo#7419, meissa/forgejo#21, forgejo/forgejo#5771
- Because we now own the fixture loader (forgejo/forgejo#7715), its
rather simple to ensure that all tables that did not receive fixtures
should be cleaned between runs and removes the need to place empty
fixture files.