fix: abuse reports string data types
While for SQLite the columns corresponding to `string` fields will have
the `TEXT` type, this is not the case for PostgreSQL and MariaDB/MySQL
(according to XORM default mapping rules _String is corresponding to
varchar(255)_).

Therefore `abuse_report`.`remarks` should be of type `VARCHAR(500)` and
`abuse_report_shadow_copy`.`raw_value` of type `LONGTEXT`.