chore(ci): add doAPICreateRepository api.CreateRepoOption option
So that Private, Template and other options can be set instead of
just AutoInit from the empty argument.

The objectFormat is kept as an argument because it needs to be
specified on most calls and it would be very verbose to require an
api.CreateRepoOption struct to be created on each occasion.