chore(ci): add doAPICreateRepository api.CreateRepoOption option
So that Private, Template and other options can be set instead of
just AutoInit from the empty argument.

The objectFormat is kept as an argument because it needs to be
specified on most calls and it would be very verbose to require an
api.CreateRepoOption struct to be created on each occasion.

(cherry picked from commit 092bbf3c3e35065e0b39f097797a8a2f9935ceb3)

Conflicts:
	tests/integration/pull_commit_test.go
	tests/integration/repo_test.go
	tests/integration/signing_git_test.go

  tests that did not exist in v11 do not need to be adapted