chore(activitypub): add more descriptive errors and warnings
This includes a series of (very subjective) changes I wish I had when
debugging, including slightly more consistent error descriptions as
well as more context for when no debugger is used (i.e. we don't
directly know which function we are in).

This fixes a couple of warnings as well.