fix: disable Forgejo Actions notifications on re-run
Notifications are based on compariing the status of a run to a
previous run. A manully triggered re-run will change the status of an
existing run. It would be complicated to document and explain under which
circumstances multiple notifications can be sent about the same run.

Notifications being disabled in the case of a manual re-run is simple
to document and understand.

Refs forgejo/forgejo#8348