fix: cancelled or skipped runs are not failures for notifications
From the point of view of a notification, the only status which must
be considered a fail is `StatusFailure`. All others are either success
`StatusSuccess` or undetermined `StatusCancelled` and `StatusSkipped`.

Those are the only four status in which a run can be when it reaches
the `ActionRunNowDone` function because it is filtered on `IsDone`
which is only true for those.