fix: user activation with uppercase email address
- Right before the call to user email activation, the user is updated
[^1]. This causes the email to be lowered, which in turn makes the call
to activate the user activation fail (on database where collation is
case senstive, which is the recommened collation by Forgejo).
- The code in `BeforeUpdate` is quite confusing, the comment has become
slightly out of date and was reworded to reflect reality and its
purpose. The code is also slightly reworked to only lower the email for
the `AvatarEmail` field to avoid causing side-effect.
- Added unit test.
- Resolves forgejo/forgejo#8354

[^1]: https://codeberg.org/forgejo/forgejo/src/commit/4927d4ee3d53ad79fe88dfe862c33339c0a9cdc6/routers/web/auth/auth.go#L785