chore: improve reliability of webauthn e2e test
- This test is the source of many transient errors
https://codeberg.org/forgejo/forgejo/issues/8359#issuecomment-5655557
and is semi-reproducible locally. Any debugging code that is added will
result in the error no longer being reproducible, making it hard to say
why this is failing.
- It no longer seems necessary to add this `waitForURL` call as
Playwright now seems to gracefully handle the case where we want to go
to a specific page while playwright might still be navigating to another URL
that was initiated by clicking on a button - thus removing the source of
the transient error altogether.