test(cli): test `admin auth` subcommands
With `admin_auth_oauth.go` now using `authService`, we can test the
behavior of its subcommands in a similar way to how we test the
subcommands of `admin_auth_ldap_test.go`. We add a test for both the
`admin auth add-oauth` and `admin auth update-oauth` commands. We try to
test as much of the visible behavior as makes sense. There are surely
some cases that we're missing here.

Some of the tests are interesting in that they are describing the
current behavior, and the current behavior is not ideal. Some examples:
- There isn't a way to set `(auth.Source).IsActive` to `false` from
    `admin auth add-oauth` or `admin auth update-oauth`. They're just
    always active when you use `admin auth add-oauth`, and you have to
    use some other process (like the UI) to deactivate them.
- There isn't a way to update `(oauth2.Source).SkipLocalTwoFA` from
    `admin auth update-oauth`. Whatever it set initially with
    `admin auth add-oauth` is all you can do from the CLI, and you have
    to use some other process (like the UI) to deactivate them.
- Both `admin auth add-oauth` and `admin auth update-oauth` allow
    setting `--custom-*` flags for providers that don't respect those
    values. E.g. Nothing prevents passing `--provider=openidConnect`,
    `--use-custom-urls=''`, and `--custom-tenant-id=foo`.
- The `--use-custom-urls` flag is a `cli.StringFlag`, when it likely
    should be a `cli.BoolFlag`. Its value is ignored completely, but it
    must be supplied a value. More to the point, if the value is
    supplied as `--use-custom-urls=false`, it's ignored because it's
    only checked for existence, not whether it is one of two valid
    boolean values.

There are other behaivors that are not ideal. But this change is only
about adding tests, not about fixing these unideal behaviors. Fixing
those behaviors are likely considered breaking changes (since they would
almost surely break someone that was unaware of the implementation and
using these flags in a way that wasn't intentional). So we're going to
leave that to a separate change.