fix: Visually distinguish the branch name in action description
In the dashboard, when someone deletes a branch, the message
that is displayed doesn't distinguish the branch name from the rest
of the message visually. Because that's user-provided content, I find
it odd, as it blends in with the system-generated message.

For consistency with other messages (such as the one obtained when
renaming a repository), I propose to show the branch name in monospaced
font via the <code></code> markup.