fix: enable multi-line math equations in wiki
- When math equation support was added into Gitea it allowed for math
equations to be typed over multiple lines via `\[ ... \]` and `$$ ...
$$`. Specifically the former delimaters caused problems with legitimate
markdown input to be seen as a math equation and therefore was disabled
in e1a82a15d3837a51409b128d4cf66d86e63788e0.
- Enable this multi-line parsing for wiki as it's less likely to cause
issues in the context of the wiki.
- It is hard to fix this issue in a proper way without investing a good
amount of time in Goldmark as explained in https://codeberg.org/forgejo/forgejo/issues/6902#issuecomment-2845317
- Added unit test.
- Resolves forgejo/forgejo#6902