refactor(cli): move `authService`
We want to use `authService` from `admin_auth_oauth.go` as well as
`admin_auth_ldap.go`. Even though all three of these files
(`admin_auth.go`, `admin_auth_ldap.go`, `admin_auth_oauth.go`) are
siblings from Go's point of view, `admin_auth.go` is the semantic
"start" of `admin auth` subcommands.  So it seems like having
`authService` in `admin_auth.go` is the more general place to put it.

`admin_auth.go` might not be the best place for `authService`. But it
seems better than using `authService` in `admin_auth_oauth.go` when it
is defined in `admin_auth_ldap.go`. If we find a better place for
`authService`, we can move it there instead.

This should be a genuine refactor:
- The external interface of `cmd` hasn't changed, only a shuffling of
    where the code in the internals of the `cmd` package is defined.