refactor(cli): use `authService` for `admin oauth`
Now that we've got `authService` in `admin_auth.go`, we move the
subcommands in `admin_auth_oauth.go` to using it. We replace the
standalone `func`s with corresponding methods on `authService`. We
rename the methods so they align with how the methods are named in
`admin_auth_ldap.go`.

The only bit that's really different from what's used in
`admin_auth_ldap.go` is that we're not using
`(authService).getAuthSource`. That does an extra check to validate the
`(auth.Source).Type`, and that could technically be a breaking change.
We want this change to only be a refactor so as to not introduce
different behavior.

This should be a genuine refactor:
- The previous `runAddOauth`/`runUpdateOauth` `func`s were not exported
    from the `cmd` package, so nothing external to the `cmd` package can
    be affected by those previous `func`s disappearing from the
    internals of the `cmd` package.
- The `authService` returned from `newAuthService` uses the exact same
    calls to `cmd.initDB`, `auth.CreateSource`, `auth.GetSourceByID`,
    and `auth.UpdateSource` as before; there is just an extra layer of
    indirection (since it goes through `authService`'s methods).