feat: use `git-diff-tree` for `checkIfPRContentChanged`
As explained in the code comment, this is in most cases (specifically
larger diffs) a faster way of comparing if two diffs are equal by
comparing them on a tree-level instead of comparing individual files.

Although I am not 100% sure this also might be a more correct method
over a three-way diff as I could not find documentation about how
three-way diffs behave via git-diff(1) and might not do the same
operation as is done in the new approach. This new approach
is (comparing against base branch via the merge-base) really the same
as how diffs for the web UI are computed.

It is now tied to the testPatchCtx but that is fixed in the next commit.