chore: disable mismatched root URL e2e test for safari
- I cannot give a good rationale why Mobile Safari and webkit is failing
this test quite consistently and also succeed quite often. We add a
script to ensure that we get a mismatched URL - but these two browsers
either ignore this script sometime or delays the execution until after
the root URL check is done.
- Because it is very hard to run webkit and mobile safari without
running a heavily containerized enviroment (without a graphical
interface) it is near impossible to debug this issue properly save
ourselves a headache and disable it instead. I find it more likely to be
a problem with my playwright test than it to be a problem with the
mismatched root URL code.
- Ref forgejo/forgejo#8359