chore: do not navigate to same URL in E2E test
- In `test.beforeEach` the browser is navigated to
`/user2/repo1/pulls/5` and the test environment is prepared by making
sure the pull request is not WIP. At the start of `simple toggle` the
browser is navigated to the same URL (and the browser is already on this
URL) which the Mobile Chrome browser does not handle gracefully and
reliabily errors with `net::ERR_ABORTED`.
- Because we are already at that URL, do not try to navigate to it
again.
- I cannot offer a rationale why the Mobile Chrome browser does give a
error when this happens in subsequent tests of 'Pull: Toggle WIP'.
- Ref forgejo/forgejo#8359