fix: use parent context for new transactions
- Older code (154 places to be exact) that want to do transactions uses
`TxContext`. If the context is not already in a transaction then it uses
`DefaultContext` for the new transaction.
- Not reusing the context it was given leads to two problems: for
tracing (forgejo/forgejo#6470) any SQL code that is executed inside this
transaction is not shown in the trace. If the context is cancelled then
the transaction is not cancelled and will always complete (given there's
no SQL error).
- When this function was introduced it didn't take a parent context,
hence the usage of `DefaultContext`. When `parentCtx` was introduced it
was only used to allow to avoid nested transactions and use the parent's
transaction. I do not see any reasons why the parent context cannot be
reused, it is reused in the newer transaction function `WithTx` without
any known problems.