feat(ui): make 'Reference' in issue sidebar more uniform
This breaks the reference's contents into a new line to make more
space for the contents, and because using a single column to present
information is something we do not do in the sidebar. Although further
changes to other components may be required so that the sidebar looks
uniform in its entirety, this small tweak is a small start.

Signed-off-by: Panagiotis "Ivory" Vasilopoulos <git@n0toose.net>