feat(ui): hide 'not set's in issue sidebar if dropdown visible
Showing this to users without the ability to add 'Due date's or
'Dependencies' makes sense (instead of displaying an empty gap in its
place), but not the other way around.

If the user is in a position to add a due date/dependencies, it makes
sense to just show the dropdown menu instead of telling them that no
such things exist for a given issue/PR. This is what this change
introduces, so as to reduce noise.

Signed-off-by: Panagiotis "Ivory" Vasilopoulos <git@n0toose.net>