feat: enable Forgejo Actions schema validation in the UI
It is disabled otherwise in contexts where a workflow schema
validation cannot be conveniently acted upon or has already been
previously validated.

It may end up with a workflow sent to the runner although it cannot
be validated. In which case the runner is responsible for running the
validation and reporting any error it finds.

Refs https://code.forgejo.org/forgejo/runner/src/branch/main/RELEASE-NOTES.md#8-0-0