fix: correct image source for quoted reply
- Since v10 replys are generated on the fly to handle quoted
reply (forgejo/forgejo#5677), this means that we have to do some work to
construct markdown that is equivalent to the HTML of the comment.
- Images are slightly strange in the context of issues and pull
requests, as Forgejo will render them in the context of the repository
and as such links such as `/attachments` become
`/user/repo/attachments`, the quoted reply did not take into account and
would use `/user/repo/attachments` as link which means it gets
transformed to `/user/repo//user/repo/attachments`.
- Instead of fixing this on the backend (and maybe break some existing
links), teach the quoted reply about this context and remove it from the
image source.

Reported-by: mrwusel (via Matrix)