Wrap delete process in a if
The method GetResolvedReports returns reports
that have been resolved.

If a duration for how long reports should be kept is passed,
then we check resolved reports for whether they are older than
the duration and return those. Otherwise all resolved reports
are returned.

We've wrapped the delete process in RemoveResolvedReports
in an if statement to keep the behaviour of GetResolvedReports
such that it can apply to either use case described above.