feat: show CI status on force-pushes
If a change is part of a force-push and the
commit(s) have a CI status, this will now be shown
after the hashes.

`interactiveBorder` has been lowered as it was
possible to activate the hover state for both
commits. It would be unreasonable to test this
within Playwright and thus this needs to be
manually tested. On a pull request page that
contains a force-push you will notice:
a) the (de)activation area for force-pushes is now
smaller, and;
b) it is not possible to activate both the hover
state/popup.

ExecuteTemplate function from @gusted
https://codeberg.org/forgejo/forgejo/pulls/5168/
https://codeberg.org/forgejo/forgejo/pulls/2884/

Close #4932