feat: add migration
This migration closely follows the migration that was done for the
keying migration for the `two_factor` table:

1. If we cannot decrypt an existing secret, we assume it's corrupted and
delete it and do not let the migration fail because of that.
2. Do a column type migration for all supported databases.

We additionally add a transaction over the whole migration so it either
happens or it doesn't happen. If there's a error, you can always fix the
database and re-run the migration without having some state left by the
previous migration. This transaction is tested, modifying it slightly to
have a deadlock will result in the migration test being
deadlocked (testd on SQLite).