chore: update broken git hook error
- The "Git hooks of this repository seem to be broken" error is
currently shown if for whatever reason the database doesn't agree with
Git on what the commitID for a branch is. This can happen for a variety
of reasons, but usually is just a transient error.
- Saying that the git hooks of the repository "looks broken" is very
likely to be correct in some contextes, but I have not yet seen that
case and in those cases; other stuff would also be broken and would
indicate that Forgejo is not functioning correctly.
- Replace this warning with a more helpful message that contains a easy
fix that works in the cases where it was a transient error - push
another commit (so the git hooks get run again). If the warning persists
contact the instance admins and they will figure out the git hooks are
broken or contact Forgejo about this problem.
- Moreover removes a link to Gitea documentation.
- A potential improvement (if people will run into this problem, and we
then know what they ran into), we could show a specific warning to
instance admins that contains troubleshooting steps or point to our own
documentation.