chore: remove goroutine PID logging
- It was possible for Forgejo adminstrators to configure the logger to
log the goroutine PID that made the `Log` call.
- The need for the goroutine PID to be logged is that it might give
insight which request or otherwise process made the log call by
correlating the PID with other logs. However even for a single request I
cannot make correlations between the goroutine PIDs and it seems that
this particular need cannot be fulfilled by the current implementation.
- The gathering of this PID is discouraged,
https://go.dev/doc/faq#no_goroutine_id, and the current implementation
is a hack and can break in future Go releases; it has broken before in
61e21d7ded198653a91a49f945d2c2a6dee0e73c.
- If the need arise, we instead should make the logger implementation
context aware and use a PID that's associated with the context, which
is guarantees to be consistent (this is also the more idiomatic way to
achieve this functionality).