feat: use native dialog for modal
- The current implementation for modals is provided by fomantic UI.
- This patch introduces a new implementation that relies on the
`<dialog>` element to provide modal, whereby the heavy lifting is done
by the browser.
- This implementation is considerably simpler, accessible (although
untested) and lightweight. It is capable of replacing fomantic UI's
modal implementation + our dimmer implementation (~2k lines of code and
CSS).[^1] As a first step the empty content modal is migrated.
- This brings in the CSS needed to display `<dialog>` and a helper
function that hides some boilerplate code that's needed to show
`<dialog>` as a modal.

[^1]: The heavy work has already been done by me in a local branch, but
reviewing that gigantic patch in one PR is not doable.