chore: deduplicate pull request template path names
I want to add a new directory to search through, and this list contains
a lot of duplication. If I want to add a "docs/" prefix, I'd be adding
it in 2 capitalizations for 3 extensions. I think it makes sense to
deduplicate, though it does make these names less greppable. To keep at
least some of that, I added one full path as an example in the comment.