fix: compare week as numbers and not as strings
The repository contributors graph received the stats for each author for
each week, these weeks are stored as unix milis values, `Object.entries`
converted these values to strings and `sort()` would thus sort them as
strings - this worked without a problem for most repository.

If a repository has commits from before 'Sun Sep  9 03:46:40 AM CEST
2001', it meant that the weeks when those commits were made would be
sorted towards the end because "1000000000" > "999999999" (when compared
as strings) and would thus be silently cut from the data. This was
edge-case was seen by the curl repository (https://mastodon.social/@bagder/115018271785548165)

Sort them as numbers to avoid this problem, them being stored as strings
is otherwise not a problem.

Co-authored-by: viceice