feat: artifacts can be downloaded using their id instead of their name
The web endpoint

/{owner}/{repo}/actions/runs/{run_id}/artifacts/{artifact_name_or_id}

can be used with either the artifact name used when it is uploaded or
the instance wide unique number of the artifact, if it is not
found. For instance:

/root/myrepo/actions/run/3/artifacts/my_artifact_name

or

/root/myrepo/actions/run/3/artifacts/42

The upload-artifact@v4 output value `artifact-url` is built in this
way and is now a valid URL to access the artifact.

Refs https://code.forgejo.org/forgejo/runner/issues/187
Refs https://code.forgejo.org/forgejo/upload-artifact/src/tag/v4#outputs