fix: don't allow credentials in migrate/push mirror URL
Do not allow credentials to be present in the URLs that are provided for
migrations and push mirrors. They have to be given via the dedicated
input fields. Give a error when this happens.

There's nothing wrong with trying have the backend "correct" this, but
would be a larger patch than necessary in the context of a security fix.
This can be done in public.

(cherry picked from commit 9f955b300b11c64f79ef6a77c92fd6e1226eeb5a)

Conflicts:
	options/locale_next/locale_en-US.json
  trivial context conflict