fix: Forgejo Actions RUN_ID is equal to RUN_NUMBER
It is now the same as `RUN_NUMBER` which is relative to the repository.

The value of `RUN_ID` in GitHub Actions is relative to the repository
and is used by upload-artifact@v4 to build the URL to download an
artifact.

The `RUN_ID` sent by Forgejo is the instance wide unique id assigned to
a job. It cannot be used for to access as it is never exposed by an
API or a web endpoint.

Refs forgejo/runner#187
Refs https://code.forgejo.org/forgejo/upload-artifact/src/tag/v4#outputs
Refs https://docs.github.com/en/actions/reference/workflows-and-actions/contexts#github-context