fix: Unreliable afterEach hook + edit btn locator
Refactor:
Hooks needs to be presented before the tests, this is a practice
for readability on what is setup for the tests and more maintainable
to change the hook when needed.

Fixes:
On the afterEach hook, we were deleting the rules and waiting for
the page load, with await page.waitForLoadState('domcontentloaded')
which behaves unreliably or being flaky on multiple runs.

In order to fix this, we can simply go for the selector count
expectation to be 0 that is not present. In this case, the delete
button should not be present after the rule is successfully deleted.

Also, where necessary - try to page.locator with selectors instead
of just page.getByText, this will lead to problems when you have
the text present elsewhere in the same page.