test(e2e): Fix and refactor repo-settings flaky test
Refactor:
---
Hooks should be placed before the tests such that it is obvious that you read what is set in hooks before reading the tests. Here, test.afterEach is placed at the top of test describe.

Uncomment the Mobile Safari browser, it has to work and is working.

Fixes:
---
On the afterEach hook, we were deleting the rules and waiting for the page load, with await page.waitForLoadState('domcontentloaded') which behaves unreliably or being flaky on multiple runs.

In order to fix this, we can simply go for the selector count expectation to be 0 that is not present. In this case, the delete button should not be present after the rule is successfully deleted.

Also, avoid using page.locator with selectors instead of just page.getByText, this will lead to problems when you have the text present elsewhere in the same page.

Give 5 secs timeout(default is 3) after the state change and when we immediately wait for the selectors to check for ContainText and HaveCount. With the current CI setup the 5 secs timeout works without failures.