feat: Remove waitFor event for Edit Button
The reason for removal of waitFor is that we are already waiting for the
page to wait for `domcontentloaded`, i.e.
await page.waitForLoadState('domcontentloaded');

The initial purpose of waitFor implemented is to find the locator of the
edit button and wait for the dom event to load. But this already gets
satisfied by the page.waitForLoadState - hence this is removed.

NOTE: Setting a timeout for the locator is a bit hacky. We have already
setup the navigation timeout in the playwright.config.ts file. If that
does not work, we will document the change on why is it happening prior
changing in the e2e codebase.