Show repo slug for rendered links to external commits
A link to the commit view is rendered into a short form that only shows
the commit SHA. To show when the link refers to an "external"
repository, i.e. one that is different to the one the user is currently
viewing, add the `<owner>/<repo>` slug like so:

`forgejo/forgejo@abcdef1234`

For the common case, parsing the repo slug is relatively simple.
However, if the Forgejo instance is hosted at a sub path, and the repo
happens to be named like a known app route, the displayed slug may not
be correct.
The underlying link will still work.

Closes: #9130