fix: ignore existence of commits for force pushes
Because we wish to show the status of the old and new commit of a force
push, ignore that the commit doesn't exist and return a commit with only
its ID filled. This is enough to still show the CI status of this commit
although the commit itself is no longer reachable.