fix: use correct commit when fetching patch view for review comments
This is effectively a clean revert of e7a77d32cc96692f5dbbd3ed1cfa7ff4a9ba026b,
which for some reason broke this logic.
Yes, always having the latest code shown in the diff is nice, but doing
this mismatches the commitID and the line number the comment is being
posted on, which very easily and frequently leads to reviews that show a
completely nonsensical patch above them, cause the code it's referring
too shifted by several lines since the time the review was started and
the time it was posted.