feat: make upload URL match Github API
Because this field was added purely to match the Github API, it is fair
to actually make it compatible with how the Github API intended it to be
and that is by adding `{?name,label}` which is used by Github's Oktokit.

We only add `{?name,label}` when we know the request accepts a Github
response. This avoids breaking the API compatability with non-Github API
clients.

Resolves Codeberg/Community#2132

c