chore: Add `AcceptsGithubResponse`
Add a function to the API router context struct to check if the requests
accepts a Github response. Although Forgejo API will never be compatible
with the Github API, historically our API has been designed to follow
that of Github closely and we know that a lot of tooling that uses the
Github API can be used against the Forgejo API with little to no
problem.

As a meet in the middle solution, this function can be used to respond
with a more appropiate response that follows the Github API. This allows
Forgejo to avoid breaking compatability with existing users of the API
and allows the API to be oh so slightly more compatible with that of
Github for API clients that expect a Github response.