WIP: support for adding dependencies by reference
This compiles now, but is untested

Reasoning for decisions so far:

> Sends the full input text instead of doing a separate search fetch
> request to find the id based on the input, before submitting the form.

If I paste a reference in and hit enter quickly, there might not be time
for the search to return, and I'd still like to have it submit.

> Prioritizes search text over issue id

The current behavior is that the rawSearch text gets cleared when
clicking out of the text box. So if there is any rawSearch text, that is
showing to the user in front of any issue that may have been previously
selected.