feat: avoid expensive SQL for org home
This change comes from Codeberg's Forgejo fork and has been in deployed
in production for a month now.

What we noticed on Codeberg is that navigating to the organisation's
home was quite slow, for example https://codeberg.org/forgejo, this was
caused by searching for the repositories that is owned by the
organisation. The expensive part of this SQL was that it tried to show
repositories where the organisation is a collaborator of... which is
none as you cannot add organisations as a collaborator. Not doing this
made the SQL query very fast again.

There's no test associated with this, as it's essentially a no-op before
and after.