chore: add fault injector hook in unittest
Forgejo has many helper functions to test various aspects of the
database state, however one aspect it is not able to test is
transactions. As this would require that some random SQL query fails to
indeed observe that the whole transaction is being rollbacked.

So how do we make a random SQL query fail? Via a fault injector hook,
which is always added to xorm (during unittest) and can be enabled on
demand to say after how many SQL queries it should start returning a
error (fault injecting).

This allows a test to do the following: after two SQL queries lets call
it a day and then execute the function that starts a transaction and
does a few SQL query. It can then observe that indeed the function was
fault injected (`ErrFaultInjected` is returned) and after querying the
database it can observe that nothing was changed and thus can conclude
the transaction was rollbacked.