chore: add unit test
To demonstrate how the fault injector test helper can be used, lets add
a test to a function I really wanted to test but couldn't because the
fault injector didn't exist. `NewTwoFactor` was recently made into a
transaction (a8c61532d29d31e804b40c44c9984d987fad8682) and although it
would not be catastrophic it would be really bad if records were being
inserted if for some reason setting the secret failed.

The test that's added demonstrates that the function uses a transaction
and rollbacks correctly.

Weirdly enough the fault injector can be viewed as testing a
specification, because it assumes nothing about how the function does
it (and you could even design a function that purposly doesn't work but
succeeds this test), it merely assumes there's a transaction and within
that transaction some SQL queries will be done. However it also needs a
certain amount of knowledge about how the function is implemented
because the developer needs to tell after how many SQL queries you want
to inject a fault and you want to do at a point where there's already a
observeable change happening in the transaction and not fault inject
if the transaction only contains `SELECT` queries.

I'm sure you could design a smart fault injector that can do such guess
work (although it sounds like a topic for a PhD thesis) and you could
design a helper function that can then guide the fault injector to find
every interesting place to do a fault injection and ensure the
transaction always falls back; as a first prototype having the
programmer tell after how many SQL queries a fault should be injected is
sufficient for a lot of the transaction we are going to test in Forgejo.